+++ src/FileParser.ts
@@ -1,78 +1,59 @@
+import RegexMatchFormatError from './exceptions/RegexMatchFormatError';
+import RegexTest from './RegexTest';
+
 export const TEST_AREA_DELIMITER = '---';
 
-const REQUIRED_FLAG = 'd';
-const DEFAULT_FLAGS = ['g', 'm'];
-
-export interface ParsedRegexTest {
-  matchingRegex: RegExp;
-  testLines: string[];
-  startTestIndex: number;
-}
+const NEW_LINE_LENGTH = 1;
 
 class FileParser {
-  static parseFileContent(fileContent: string): ParsedRegexTest | undefined {
+  static parseFileContent(fileContent: string) {
     const fileLines = fileContent.split('\n');
 
-    const firstDelimiterLineIndex = this.findFirstDelimiterLineIndex(fileLines);
-    if (
-      firstDelimiterLineIndex === undefined ||
-      firstDelimiterLineIndex === fileLines.length - 1 ||
-      firstDelimiterLineIndex === 0
-    ) {
-      return undefined;
-    }
+    const regexTests: RegexTest[] = [];
+    let regexLineIndex: number | undefined;
+    let testLines: string[] = [];
+    let isInTestStringArea = false;
 
-    const regexLine = fileLines[firstDelimiterLineIndex - 1];
-    const matchingRegex = this.transformStringToRegExp(regexLine);
+    let startTestIndex = 0;
+    let charCount = 0;
 
-    if (!matchingRegex) {
-      return undefined;
-    }
-
-    const testLines = this.getTestLines(fileLines.slice(firstDelimiterLineIndex), matchingRegex.multiline);
-
-    const startTestIndex = fileContent.indexOf(TEST_AREA_DELIMITER) + TEST_AREA_DELIMITER.length + 1;
-
-    return { matchingRegex, testLines, startTestIndex };
-  }
-
-  static findFirstDelimiterLineIndex(fileLines: string[]): number | undefined {
     for (let i = 0; i < fileLines.length; i++) {
       if (this.isTestAreaDelimiter(fileLines[i])) {
-        return i;
+        isInTestStringArea = !isInTestStringArea;
+
+        if (isInTestStringArea) {
+          if (regexLineIndex === undefined) {
+            throw new RegexMatchFormatError(i);
+          }
+
+          startTestIndex = charCount + fileLines[i].length + NEW_LINE_LENGTH;
+        } else if (regexLineIndex !== undefined) {
+          const regexTest = new RegexTest({
+            regexPattern: fileLines[regexLineIndex],
+            regexLineIndex,
+            testLines,
+            startTestIndex,
+          });
+
+          regexTests.push(regexTest);
+
+          testLines = [];
+          regexLineIndex = undefined;
+        }
+      } else if (isInTestStringArea) {
+        testLines.push(fileLines[i]);
+      } else if (fileLines[i] !== '') {
+        regexLineIndex = i;
       }
-    }
-  }
 
-  static transformStringToRegExp(patternString: string): RegExp | undefined {
-    const matchGroups = patternString.match(/^\/?(.*?)(?<flags>\/[igmsuy]*)?$/i);
-
-    if (matchGroups) {
-      const [, pattern] = matchGroups;
-      const flagsGroup = matchGroups.groups?.flags;
-
-      let flags = flagsGroup?.replace('/', '');
-
-      if (!flags) {
-        flags = DEFAULT_FLAGS.join('');
-      }
-
-      if (!flags.includes(REQUIRED_FLAG)) {
-        flags += REQUIRED_FLAG;
-      }
-
-      return new RegExp(pattern, flags);
-    }
-  }
-
-  static getTestLines(fileContent: string[], multineFlag: boolean): string[] {
-    const testLines = fileContent.filter((line) => line !== TEST_AREA_DELIMITER);
-
-    if (multineFlag) {
-      return testLines;
+      charCount += fileLines[i].length + NEW_LINE_LENGTH;
     }
 
-    return [testLines.join('\n')];
+    if (isInTestStringArea || regexLineIndex !== undefined) {
+      throw new RegexMatchFormatError(0);
+    }
+
+    return regexTests;
   }
 
   private static isTestAreaDelimiter(line: string): boolean {