+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.3.0
+
+- Enables multiple regex testing in the regex match window.
+- Correction of regex testing with `\n` and wildcard character.
+
 ## 0.2.0
 
 - Highlight capturing groups in the test string.